changed styles of header and side panels
